Cpp14.g4 - bugfix - multiline macro parse consumes all remaining input
The multi line macro has a rule to the extent of parse until you see a \\. Seeing that this has higher precedence than Directive, this causes files with # symbol to have all remaining input be consumed within the MultiLineMacro rule. Also cleaned up the \r and \n handling to more appropriately handle line endings.

Not adding any test as the existing examples should already catch this. (Wouldn't cause a parser error, but the syntax tree created is wrong)