Restrict LineTerminators in ECMAScript grammars.
According to the ECMAScript 5.1 Standard (and the comments in the existing
ANTLR grammars) LineTerminator is not allowed between a 'return', 'throw',
'break' or 'continue' keyword and the following Identifier or Expression
or between a LeftHandSideExpression and a postfix operator. However,
while the latter case is handled correctly, this restriction is missing
from the prior one. The patch fixes this.