Create python2.g4
I just wanted to see if I could pull the parser grammar off the Python site and generate a parser written in Python 2. It compiled pretty much  as-is. I only had to make very minor syntactical changes to the parser grammar.
And added in some lexer rules and handling for  INDENT's and DETENT's.