support CURRENT_TIMESTAMP default value
As described in #962, current grammar does not allow for CURRENT_TIMESTAMP keyword as a default value for TIMESTAMP and DATETIME columns. This pull request aims to add such capability. Probably the new rules need to be labeled so comments and additions are welcome. The changes still look a bit like a workaround for a specific case, but it *is* a specific case in MySQL syntax as they clearly state in their docs.