Addressing SkipChannel issue
This commit addresses the `ReferenceError: SkipChannel is not defined` I was experiencing while parsing the `examples/strings.php` file.

I have no idea why SkipChannel has issues but ErrorLexem and PhpComments channels are fine.

Here I've complete removed SkipChannel and replaced it with stright `skip` actions.

I'm not sure the original indended purpose of SkipChannel but this works for my use and I'd like to leave this decision to accept or deny this change up to the maintainers.

If this is not acceptable can someone please explain the purpose of SkipChannel? Thanks!