[JavaScript] Improve classElement parsing.
Make sure to parse static method definitions, even if strict
mode is not applied in the global scope (the content of
JavaScript classes are inside an implicit strict scope).
In this case, the 'static' word will be recognised as an
Identifier, hence it also has to be considered in classElement
parsing.