[javascript/Go] fix compile time error
The error was caused because ANTLR4 generates 2 constructors for each RuleContext in the grammar: `New[RuleName]Context` and `NewEmpty[RuleName]Context`.

This caused a conflict since the grammar defines a `statement` and `emptyStatement` rule. Thus generating function names `NewStatementContext, NewEmptyStatementContext, NewEmptyStatementContext, NewEmptyEmptyStatementContext`.

The error was fixed by simply adding an underscore (_) to the `emptyStatement` rule.