Add RunAs block
Apex has a block used in tests with the System.runAs method.
This allows portions of a test to be run under the context of a specified user.
Documentation for this method can be found here:
https://developer.salesforce.com/docs/atlas.en-us.apexcode.meta/apexcode/apex_testing_tools_runas.htm

See apex/examples/ExampleRunAs.cls for a contrived example.

I've added a new keyword called RUNAS because this method must
be used as a block. It is also invalid to use this method in a typical
fashion such as System.runAs(user); System is also reserved for class names so
there cannot be a custom class called System with a method called runAs.