Rename NULL, TRUE, and FALSE literals in SQLite grammar to successfully compile targeting c++
ANTLR runtime includes header files that define NULL via preprocessing directive, using NULL as an identifier in grammars (specifiecally the lexer) causes it to be used
as an identifier in the corresponding autogenerated c++ code which then fails to compile.

We change NULL to NULL_ in SQLiteLexer.g4 and update its use in SQLiteParser.g4. We also do the same to TRUE and FALSE to avoid similar issues that may occur when compiling on windows.

This fixes #1904