Allow to check colmun
In this case:
SELECT 1 AS toto, table.Id
FROM table
GROUP BY toto

To check column in select, we must parse first FROM clause. But then parse FromClause, in previous g4, group by clause is included in fromclause and we cannot knonw the alias in select. 
Thus, we must separe fromclause and groupby clause to simply the sql parsing.