Don't depend on Powershell Get-ChildItem returing absolute paths.
The _scripts\test.ps1 script fails all tests on my machine (Windows 11, version 21H2, OS Build 22000.978). The issue is that the test filenames returned from Get-ChildItem are relative to the parent directory on my machine. They appear to be absolute in the CI builds (see https://github.com/antlr/grammars-v4/actions/runs/3064889059/jobs/4948444850#step:17:39 for an example). I've fixed that in a manner that will work regardless of the Get-Childitem results, but it's disturbing. I can't find any documentation that states or implies the absolute pathname results.