[build] Fixes for #3468 -- Dependabot PRs cause crashes. (#3469)
* Fixes for https://github.com/antlr/grammars-v4/issues/3468

Remove all mention of tool version in workflows/main.yml, because the version is per-target. Instead install antlr4 tool.

* Fixes for #3468.

Set versions for Java Antlr version and use antlr4 tool to build.

* Fix for #3468. Cpp.

Remove all explicit checks for java and just use the antlr4-tools antlr4 tool.

* Fix for #3468. Java and install of Python3 and Pip.

Fix the Java template, and install of Python3 and Pip regardless because we now use antlr4, and never install java explicitly.

* Fixes for #3468. JavaScript.

Fix the JavaScript build scripts, including the package.json version for Antlr4 runtime.

* Fixes for $3468. TypeScript.

Fix the version number for Antlr4 tool and runtime, and grep the version from the package.json file and send that the the antlr4 tool.

* Fix #3468. JavaScript, TypeScript, Dart.

Fixes to remove newlines from brain-damaged version number that contain newlines.

* Fixes for #3468. Java.

Set version for Antlr. Set up correct location of jar file using Python script from antlr4 tool.

* Fixes for #3468. Java and Powershell.

This is a change to the Powershell version of the build script, equivalent to the Bash version.

* Fixes for #3468. Java and Powershell running.

This is a fix for the jar path for running the parser with java.

* Fixes for #3468. Java and Powershell.

Variable references require curly braces around the name when colon follows the name.

* Fixes for #3468. PHP and Bash.

Change reference to java and the antlr jar to the antlr4 tool and version.

* Fixes for #3468. Python3, CSharp.

Updates to version number of antlr for CSharp, scripts for Python3.

* Fixes for #3468. Go target.

These changes are to fix the problem of using the antlr4 tool and using the latest api for 4.13.0 for Go. It's all changed.

* Fixes for #3468. Cpp and Pwsh

Updates to get Cpp target to work. Adjustments to pwsh scripts.

* Fixes for #3468. Fix formatting of abb grammar, call all build of all targets.

This change is to the abb grammar, but the only thing that is done is to remove some triple spacing between some rules. The point is to force a rebuild of all targets, so the new build scripts can all be checked.

* Fixes for #3468. Dart target still on 4.12.0.

Rollback the version of Antlr to 4.12.0 for Dart--the runtime has not been released yet. In addition, this is a fabulous test of the new per-target Antlr version testing.

* Fixes for #3468. Dart.

An release of the Dart runtime for 4.13.0 has been made in the last day.

* Fixes for #3468. Remove the .errors file mistakenly commited.