Fixes for #3480 (#3491)
This fixes #4580, where .csproj files were not recognized and updated by Dependabot. In this implementation, file Other.csproj is added, along with Other.csproj.meta. The Other.csproj file is in MSBuild format, not StringTemplate format, as there is a .meta file that turns off StringTemplate processing, and just copies the file to the generated driver directory. The Test.csproj file now contains an "Import" statement, where it includes Other.csproj. Because there are two csproj files, the rest of the changes clear up the ambiguity with dotnet builds.