Changes withElement definition to be closer to CQL3 specification.
This PR adds support for two things:

1. Specifying order of more than one clustering keys, and
2. Allowing for additional table options after `WITH CLUSTERING ORDER`.

Both are allowed in CQL3, see the relevant section of the spec:
https://cassandra.apache.org/doc/stable/cassandra/cql/ddl.html#create-table-statement

This does not add support for `COMPACT STORAGE`, which is also allowed by the spec.
Adding it as part of this PR would greatly complicate the grammar and break backwards
compatibility. Someone would have to think more carefully about how to roll out such
a change.