Update grammar to be target agnostic.
Antlr4 "@header" blocks are target specific. But, we need this to work for any target. So, we use a comment in the grammar and transformGrammar.py to fix up the problem. _insideOptionsBlock is removed because it's not used anywhere. Untabify source code--tabs are presentation specific and thus depend on the editor used to open it. We don't use the same editors people.