[PlSql] Add support for CREATE SCHEMA (#4436)
* Update to version 0.23.15 of Trash Toolkit.

* Update dotnet-tools.json

* Update to 0.23.16, add trapc.

* Remaster due to changed grammar.

* Remove all broken ports for antlr/antlr4.

All ports need to be rewritten in target agnostic format.

* Remove testing of PHP for trapc--PHP is not a reliable runtime. Add antlr-ng generator for Antlr4ng testing.

* Update grammar to be target agnostic.

Antlr4 "@header" blocks are target specific. But, we need this to work for any target. So, we use a comment in the grammar and transformGrammar.py to fix up the problem. _insideOptionsBlock is removed because it's not used anywhere. Untabify source code--tabs are presentation specific and thus depend on the editor used to open it. We don't use the same editors people.

* Create LexerAdaptor.ts

* Create transformGrammar.py

* Update LexerAdaptor.ts

* Fix missing if-then-else.

* Add Antlr4ng port.

* Add Cpp target.

* Update to latest of Trash for Cpp testing.

* Initial version of Dart port--does not work yet.

* Fix Dart port.

* [PlSql] Add support for CREATE SCHEMA

* Add TypeScript port.

* Add Python3 port.

---------

Co-authored-by: Ken Domino <kenneth.domino@domemtech.com>