Remove one ambiguity in golang grammar.
This code removes an ambiguity between conversion and function calls. We don't have a symbol table, so we assume that if we have "ID(...)" we will assume a function call, otherwise it's a conversion. The change improves the performance by more than double for restorer.go (see #4449). In addition, there are numerous problems in the refactoring of the grammar from the official spec for golang. I'm changing it back to reflect what is actually in the spec, but will employ correct refactoring techniques in the future. The use of EOF on the RHS of a lexer rule is way, way incorrect!! The use of EOF in eos is also way, way incorrect.