[golang] Fix for #4448: remove ambiguity. (#4459)
* Remove one ambiguity in golang grammar.

This code removes an ambiguity between conversion and function calls. We don't have a symbol table, so we assume that if we have "ID(...)" we will assume a function call, otherwise it's a conversion. The change improves the performance by more than double for restorer.go (see #4449). In addition, there are numerous problems in the refactoring of the grammar from the official spec for golang. I'm changing it back to reflect what is actually in the spec, but will employ correct refactoring techniques in the future. The use of EOF on the RHS of a lexer rule is way, way incorrect!! The use of EOF in eos is also way, way incorrect.

* Fix CSharp port.

* Fix rest of the current ports.

* Fix Java port, add Antlr4ng and Dart ports for golang.

* More fixes to ports.

* More fixes to ports.

* Fix Cpp port.

* Fix for #4448

Add in predicate to make sure "chan" is not chosen if followed by "<-". Remove bad readme.md.

* Fix typos.