Fix for #4492
This is a test suite of .lua files accumlated from various links mentioned in the grammar. Careful attention was placed to check out a specific version if mentioned. Of the files added, 179 tests failed to parse, and thus renamed with the '.lua.fail'. As the grammar is fixed, these can be renamed back with the '.lua' extension. The desc.xml was updated to perform a full test for the CSharp target in order to save on valuable Github Actions testing cycles. As is, there is a 5 minute time limit for all tests, so the list may need refining.