[lua] Fix for #4492 -- add large test suite. (#4493)
* Fix for #4492

This is a test suite of .lua files accumlated from various links mentioned in the grammar. Careful attention was placed to check out a specific version if mentioned. Of the files added, 179 tests failed to parse, and thus renamed with the '.lua.fail'. As the grammar is fixed, these can be renamed back with the '.lua' extension. The desc.xml was updated to perform a full test for the CSharp target in order to save on valuable Github Actions testing cycles. As is, there is a 5 minute time limit for all tests, so the list may need refining.

* Update Trash to version 0.23.18.

* Update desc.xml.

Set up subset of total test suite in order to take care of Trash testing considerations.

* Remove iconv testing/filter of test cases as this is very slow and does not work apparently for the lua tests.

It's not clear why not all the inputs in the lua test suite are valid utf-8 according to iconv, e.g., db.lua. I am removing this filter because these files parse anyways. This requires more investigation.

* Add full check of test suite for CSharp target.

* Update pom.xml to do positive tests.