[lua] Fixes for ambiguity and tree structure. (#4502)
* Fix #4496.

Fix broken ports and add Antlr4ng port. The grammar comment is changed to reflect that the header is required not only for Cpp but for Antlr4ng.

* Fix #4499 -- Add parse tree validation via XPath predicates.

* Update to Trash 0.23.20, templates here.

The update to 0.23.20 is to use the templates that are checked in here (they are identical).

* Reformat Go code.

(Changed to push rebuild, which failed because I didn't let nuget.org have enough time to reindex toolkit.)

* Update readme.

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Update readme.md

* Remove recursion in functioncall.

* Add semantic predicate to disambiguate stat problem.

* Update to lastest for Trash toolkit.

* Fix for #3885.

* Remove ambiguity. Improve readability of refactoring steps.

* Fixed Antlr4ng, Cpp, Dart ports. Renamed non UTF-8 files to .lua.fail. Fix header comment in grammar.

* Fixed Go port.

* Fixed Java port.

* Add TypeScript port.

* Update JavaScript port.

* Update Python3 port.