Fix for #4428. (#4513)
The purpose of the grammar ResourcePlanParser.g4 is unknown. It requires `identifier`, but that's defined in another .g4. This grammar includes IdentifiersParser.g4 (or vice versa). And, no other .g4 includes ResourcePlanParser.g4 (and along with IdentifierParser.g4). I'm renaming this to have it not interfere with trgen. Otherwise, it's a "top-level" grammar, and all top-level grammars are processed by the Antlr4 tool.