[build; lua, tcpheader, trapc, plsql] Fixes for #4521 (#4523)
* Fix #4521

Fix expected number of "stat" nodes.

* Updates to templates for file encoding/binary.

* Implement binary output for parse trees for Go target.

* Update JavaScript templates for file encoding.

* Add Dart target.

* Add debugging.

For some reason, the test.ps1 is not testing both lua and tcpheader on ubuntu. Find out why.

* Fix what grammars to test.

If we change a file in "lua/examples", we should test the lua grammar because "./lua/examples/" starts with "./lua/". Likewise, we need to test because ".\lua\examples" starts with ".\lua\".

* Fix typo.

* Remove debugging output; code works now.

* Reformat.

* Updates to Cpp target templates.

* Remove accidental file.

* Fix conversion for utf8 files, including files that contain invalid chars.

* Update st.CMakeList.txt with added EncodingInputStream.cpp.

* Update to latest release of trgen for trapc/Cpp.

* Reformat templates.

* Force build of trapc with newest trgen.

* Update with file_encoding from trgen, and add Python3 target.