Add in two pass entry point.
For a two-pass parse, we create a rule that has two calls to proto. Between the two, we rewind the parser and try a second time. For the second pass, the symbol table is constructed as before, but since they are duplicates, nothing gets created and overwritten (or at least shouldn't). These changes for CSharp; Java next.