[protobuf2, 3] Fix for #4535: disambiguate grammar via symbol table (#4538)
* Fix for #4535

* Reorganized protobuf grammars.
* Add preliminary version of symbol table.
* Add preliminary SDT actions and predicates for type resolution.

* Update testing.

* Update .errors files.

These files are produced because if a field type is not declared, it cannot be either an enum or message type. Probably to restrictive.

* Fixing problems.

* Fix problem.

* Reformat grammar.

* Fix pom.xml with reorganization of protobuf.

* "required" is a protobuf2 keyword, not version 3.

* Updates.

* Add in Java port
* Add in readme.

* Add in two pass entry point.

For a two-pass parse, we create a rule that has two calls to proto. Between the two, we rewind the parser and try a second time. For the second pass, the symbol table is constructed as before, but since they are duplicates, nothing gets created and overwritten (or at least shouldn't). These changes for CSharp; Java next.

* Rename test files.

* Add in new tests.

Make sure the use of message type works the same as for enum type.

* Fix tests.

* Rename and clean up description.

* Change code to use Antlr API. Update Java port. Reformat.

* Update readme.

* Add symbol table output for verification.

* Add Google protobuf include tests, reorganize.

* Update Java port with import statements.

* Hide files that can't be opened.

* Updates for disambiguation.

This commit fixes a couple of problems:
* It removes the explicit two pass "proto" and moves one inside the action.
* Predicates are added to disallow keywords in messagteType and enumType inside type_ as the keywords are handled with the additional types.
* Bump Trash toolkit being used to fix problems with file name not being passed to SetupParse2(). Template for CSharp updated.

* Forgot Java port.

* Fix processing of message and enum types, removing ambiguity.

* Added disambiguation for bool literals vs full identifiers.

* Add Antlr4ng port. Refine readme.

* Something is wrong with TypeScript "as"-operators.

Replaced with code that tests the name of the object, then casts and calls method to get child.

* Add TypeScript port. Clean up CSharp port.

* Update readme.md